wip: add type tests/linting
